# editoast: speed up infra deletion